+++ formula-android/src/main/java/com/instacart/formula/integration/FragmentFlowRenderView.kt
@@ -7,8 +7,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
-import com.instacart.formula.RenderView
 import com.instacart.formula.Renderer
+import com.instacart.formula.RenderView
 import com.instacart.formula.fragment.BaseFormulaFragment
 import com.instacart.formula.fragment.FormulaFragment
 import com.instacart.formula.fragment.FragmentContract
@@ -132,7 +132,7 @@ internal class FragmentFlowRenderView(
         activity.supportFragmentManager.registerFragmentLifecycleCallbacks(callback, false)
     }
 
-    override val renderer: Renderer<FragmentFlowState> = Renderer.create {
+    override val render: Renderer<FragmentFlowState> = Renderer {
         updateVisibleFragments(it)
 
         fragmentState = it