+++ formula/src/test/java/com/instacart/formula/RendererTest.kt
@@ -22,9 +22,9 @@ class RendererTest {
     }
 
     @Test fun `avoid bad memoization if state update triggers another state update`() {
-        val subject = TestSubject<String> { renderer, value ->
+        val subject = TestSubject<String> { render, value ->
             if (value == "first") {
-                renderer.render("second")
+                render("second")
             }
         }
         subject.render("first")
@@ -55,9 +55,9 @@ class RendererTest {
     }
 
     @Test fun `render call triggers another render with same value`() {
-        val subject = TestSubject<String?> { renderer, value ->
+        val subject = TestSubject<String?> { render, value ->
             if (value == null) {
-                renderer.render(null)
+                render(null)
             }
         }
         subject.render(null)
@@ -66,7 +66,7 @@ class RendererTest {
 
     @Test fun `handling exceptions in rendering`() {
         var crash: Boolean = true
-        val subject = TestSubject<String?> { renderer, value ->
+        val subject = TestSubject<String?> { render, value ->
             if (crash) {
                 crash = false
                 throw IllegalStateException("you can't do this")
@@ -87,16 +87,14 @@ class RendererTest {
     class TestSubject<T>(private val postRender: (Renderer<T>, T) -> Unit = { _, _ -> Unit }) {
         private val results = mutableListOf<T>()
         lateinit var reference: Renderer<T>
-        private val renderer = Renderer.create<T> {
+        val render = Renderer.create<T> {
             results.add(it)
             postRender(reference, it)
         }
 
         init {
-            reference = renderer
+            reference = render
         }
-
-        fun render(model: T) = renderer.render(model)
 
         fun assertRenderedValues(vararg values: T) {
             assertThat(results).containsExactly(*values)